Conflicts:
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/config/Config.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/handler/ProtocolMessageHandler.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/message/HandshakeMessage.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/message/ServerHelloMessage.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/preparator/ClientHelloPreparator.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/protocol/preparator/HandshakeMessagePreparator.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/state/TlsContext.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/workflow/action/executor/ReceiveMessageHelper.java
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/workflow/action/executor/SendMessageHelper.java

+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/parser/extension/TokenBindingExtensionParserTest.java
@@ -61,7 +61,7 @@ public class TokenBindingExtensionParserTest {
         message = parser.parse();
         assertArrayEquals(extensionType.getValue(), message.getExtensionType().getValue());
         assertArrayEquals(tokenbindingVersion.getByteValue(), message.getTokenbindingVersion().getValue());
-        assertEquals(parameterLength, (int) message.getParameterListLength().getValue());
+        assertEquals(parameterLength, (long) message.getParameterListLength().getValue());
         assertArrayEquals(keyParameter, message.getTokenbindingKeyParameters().getValue());
         assertTrue(extensionLength == message.getExtensionLength().getValue());
     }
